Moved saga and slice of popular movies list into MoviesListPage feature
