+++ src/App.js
@@ -3,7 +3,7 @@ import Navigation from "./Navigation";
 import { toMoviesList, toActorsList, toPerson, toMovie } from "./routes";
 import ActorsList from "./features/ActorList";
 import { ActorsData } from "./features/ActorPage";
-import MoviesListPage from "../src/features/MoviesListPage";
+import MoviesListPage from "./features/MoviesListPage/components";
 import { MovieDetailsPage } from "./features/MovieDetailsPage/components/index";
 
 function App() {
